# Make generic blas/lapack routines allocation-free