--- a/src/Utilities/GenericBlasLapackFragments.jl
+++ b/src/Utilities/GenericBlasLapackFragments.jl
@@ -33,13 +33,16 @@ Base.setindex!(A::StridedReshape,v,i,j)= @inbounds A.v[idx(A,i,j)]=v
 # Reshape matrix with leading dimension lda>=m, taking into account the
 # (for standard blas entirely legal)
 # possibility that for the largest column only m elements are stored (instead of lda)
-function strided_reshape(A,lda,m,n)
+
+
+
+function strided_reshape(A::AbstractVector{T},lda,m,n)::Union{StridedReshape{T},SubArray{T, 2, Base.ReshapedArray{T, 2, SubArray{T, 1, Vector{T}, Tuple{UnitRange{Int64}}, true}, Tuple{}}, Tuple{UnitRange{Int64}, UnitRange{Int64}}, false},Base.ReshapedArray{T, 2, SubArray{T, 1, Vector{T}, Tuple{UnitRange{Int64}}, true}, Tuple{}}} where T
     if lda == m
         #
         # In this case we can assume that the A buffer is large enough to hold
         # all elements of the reshaped matrix:
         #
-        reshape(view(A,1:lda*n),m,n)
+        x=reshape(view(A,1:lda*n),m,n)
     else
         if length(A)>=lda*n
             #
@@ -69,21 +72,20 @@ end
 # - No need to return LU object 
 # - Remove unused parameters - always do pivoting anyway
 #
-# Needed only for the rare (?) StridedReshape case.
 #
-function glu!(A::StridedReshape{T}, ipiv) where T
-    # Extract values
-    m, n = size(A)
+function ggetrf!(m,n,A::AbstractVector{FT},lda,ipiv) where FT
+
     minmn = min(m,n)
-    
+    idx(i,j)=(j-1)*lda+i
+
     @inbounds begin
         for k = 1:minmn
             # find index max
             kp = k
             if k < m #   pivot === RowMaximum() &&
-                amax = abs(A[k, k])
+                amax = abs(A[idx(k,k)])
                 for i = k+1:m
-                    absi = abs(A[i,k])
+                    absi = abs(A[idx(i,k)])
                     if absi > amax
                         kp = i
                         amax = absi
@@ -91,130 +93,392 @@ function glu!(A::StridedReshape{T}, ipiv) where T
                 end
             end
             ipiv[k] = kp
-            if !iszero(A[kp,k])
+            if !iszero(A[idx(kp,k)])
                 if k != kp
                     # Interchange
                     for i = 1:n
-                        tmp = A[k,i]
-                        A[k,i] = A[kp,i]
-                        A[kp,i] = tmp
+                        tmp = A[idx(k,i)]
+                        A[idx(k,i)] = A[idx(kp,i)]
+                        A[idx(kp,i)] = tmp
                     end
                 end
                 # Scale first column
-                Akkinv = inv(A[k,k])
+                Akkinv = inv(A[idx(k,k)])
                 for i = k+1:m
-                    A[i,k] *= Akkinv
+                    A[idx(i,k)] *= Akkinv
                 end
             end
             # Update the rest
             for j = k+1:n
                 for i = k+1:m
-                    A[i,j] -= A[i,k]*A[k,j]
+                    A[idx(i,j)] -= A[idx(i,k)]*A[idx(k,j)]
+                end
+            end
+        end
+    end
+    0
+end
+
+
+
+
+#
+# C=alpha*transA(A)*transB(B) + beta*C
+#
+function ggemm!(transA,transB,m,n,k,alpha,A::AbstractVector{T},lda,B::AbstractVector{T},ldb,beta,C::AbstractVector{T},ldc) where T
+    oneT=one(T)
+    zeroT=zero(T)
+    nota= (transA=='n')
+    notb= (transB=='n')
+
+
+    aidx(i,j)=(j-1)*lda+i
+    bidx(i,j)=(j-1)*ldb+i
+    cidx(i,j)=(j-1)*ldc+i
+
+    
+    # skip the iput tests
+
+    
+    if m==0 || n==0 || ( (iszero(alpha) || k==0) && isone(beta))
+        return
+    end
+    
+    if iszero(alpha)
+        if iszero(beta)
+            @inbounds for i=1:n*m
+                C[i]=zeroT
+            end
+        else
+            @inbounds for i=1:n*m
+                C[i]*=beta*C[i]
+            end
+        end
+        return 
+    end
+    
+    if notb
+        if nota
+            #   @Inbounds Form  C := alpha*A*B + beta*C.
+            @inbounds for j = 1:n # DO 90
+                if iszero(beta)
+                    @inbounds for i=1:m
+                        C[cidx(i,j)] = zeroT
+                    end
+                elseif !isone(beta)
+                    @inbounds for i=1:m
+                        C[cidx(i,j)] *= beta
+                    end
+                end
+                @inbounds for l=1:k
+                    temp = alpha*B[bidx(l,j)]
+                    @inbounds for i=1:m
+                        C[cidx(i,j)] += temp*A[aidx(i,l)]
+                    end
+                end
+            end
+        else
+            #  @Inbounds Form  C := alpha*A**T*B + beta*C
+            @inbounds for j=1:n
+                @inbounds for i=1:m
+                    temp=zeroT
+                    @inbounds for l=1:k
+                        temp += A[aidx(l,i)]*B[bidx(l,j)]
+                    end
+                    if beta==zeroT
+                        @inbounds   C[cidx(i,j)] = alpha*temp
+                    else
+                        @inbounds  C[cidx(i,j)] = alpha*temp + beta*C[cidx(i,j)]
+                    end
+                end
+            end
+        end
+    else
+        if nota
+            # @Inbounds Form  C := alpha*A*B**T + beta*C
+            @inbounds for j=1:n
+                if iszero(beta)
+                    @inbounds for i=1:m
+                        C[cidx(i,j)] = zeroT
+                    end
+                elseif !isone(beta)
+                    @inbounds for i=1:m
+                        C[cidx(i,j)] = beta*C[cidx(i,j)]
+                    end
+                end
+                @inbounds for l=1:k
+                    temp = alpha*B[bidx(j,l)]
+                    @inbounds for i=1:m
+                        C[cidx(i,j)] += temp*A[aidx(i,l)]
+                    end
+                end
+            end
+        else
+            #  @Inbounds Form  C := alpha*A**T*B**T + beta*C
+            @inbounds for j=1:n
+                @inbounds for i=1:m
+                    temp=zeroT
+                    @inbounds for l=1:k
+                        temp +=  A[aidx(l,i)]*B[bidx(j,l)]
+                    end
+                    if iszero(beta)
+                        @inbounds   C[cidx(i,j)] = alpha*temp
+                    else
+                        @inbounds C[cidx(i,j)] = alpha*temp + beta*C[cidx(i,j)]
+                    end
                 end
             end
         end
     end
 end
 
-#
-# In the general case we can conveniently fall back to the standard Julia implementation
-#
-function glu!(A,  ipiv)  where T
-    n=size(A,2)
-    ipiv[1:n].=lu!(A).p
-    return 0
-end
-
-
-#
-# C=alpha*transA(A)*transB(B) + beta*C
-#
-function ggemm!(transA,transB,m,n,k,alpha,A,lda,B,ldb,beta,C,ldc)
-    rA= transA=='n' ? strided_reshape(A,lda,m,k) :  transpose(strided_reshape(A,lda,k,m))
-    rB= transB=='n' ? strided_reshape(B,ldb,k,n) :  transpose(strided_reshape(B,ldb,n,k))
-    rC= strided_reshape(C,ldc,m,n)
-    mul!(rC,rA,rB,alpha, beta)
-    true
-end
 
 
 #
 # Y=alpha*transA(A)*X + beta*Y
 #
-function ggemv!(transA,m,n,alpha,A,lda,X,beta,Y)
+function ggemv!(transA,m,n,alpha,A::AbstractVector{T},lda,X,beta,Y) where T
     if m==0 || n==0
         return
     end
-    rA=strided_reshape(A,lda,m,n)
+    
     if transA=='n'
-       @views mul!(Y[1:m],rA,X[1:n],alpha, beta)
+        @inbounds for i=1:m
+            Y[i]*=beta
+        end
+        ii0=1
+        @inbounds for j=1:n # DO 60
+            ii=ii0
+            alphax=alpha*X[j]
+            @inbounds for i=1:m # DO 50
+                Y[i]+=alphax*A[ii]
+                ii+=1
+            end
+            ii0+=lda
+        end
     else
-       @views mul!(Y[1:n],transpose(rA),X[1:m],alpha, beta)
+        ii0=1
+        @inbounds for j=1:n # DO 120
+            Y[j]*=beta
+            temp=zero(T)
+            ii=ii0
+            for i=1:m # DO 110
+                temp+=A[ii]*X[i]
+                ii+=1
+            end
+            Y[j]+=alpha*temp
+            ii0+=lda
+        end
     end
     true
 end
 
-#
-# In-place LU factorization of A
-#
-function ggetrf!(m,n,A::AbstractVector{FT},lda,ipiv) where FT
-    glu!(strided_reshape(A,lda,m,n),ipiv)
-    return 0
-end
+
 
 
 
 #
 # Triangular solve
 #
-function gtrsm!(side,uplo,transA,diag, m,n,alpha,A, lda, B, ldb)
-    
+function gtrsm!(side,uplo,transA,diag, m,n,alpha,A::AbstractVector{T}, lda, B::AbstractVector{T}, ldb) where T
+
     if m==0 || n==0
         return
     end
     
-    k= side=='l' ? m : n
-
-    rA=strided_reshape(A,lda,k,k)
-
-    if diag=='n'
-        if uplo=='u'
-            tA=UpperTriangular(rA)
-        else
-            tA=LowerTriangular(rA)
+    aidx(i,j)=(j-1)*lda+i
+    bidx(i,j)=(j-1)*ldb+i
+    
+    oneT=one(T)
+    zeroT=zero(T)
+    
+    lside = (side=='l')
+    nounit = (diag=='n')
+    upper = (uplo=='u')
+    
+    # skip input check
+    
+    if iszero(alpha)
+        @inbounds for i=1:m*n
+            B[i]=zeroT
         end
-    else
-        if uplo=='u'
-            tA=UnitUpperTriangular(rA)
-        else
-            tA=UnitLowerTriangular(rA)
-        end
+        return
     end
     
-    
-    if transA=='t'
-        kA=transpose(tA)
-    else
-        kA=tA
-    end
-    
-    
-    rB=strided_reshape(B,ldb,m,n)
-    
-    if  side=='l'
-        rB.=kA\(alpha*rB)
-    else
-        rB.=alpha*rB/kA
+    if lside
+        if transA== 'n'
+            # @Inbounds Form  B := alpha*inv( A )*B.
+            if upper
+                @inbounds for j=1:n
+                    if !isone(alpha)
+                        @inbounds for i=1:m
+                            B[bidx(i,j)] *= alpha
+                        end
+                    end
+                    @inbounds for k=m:-1:1
+                        if !iszero(B[bidx(k,j)])
+                            if nounit
+                                B[bidx(k,j)] /= A[aidx(k,k)]
+                            end
+                            @inbounds for i=1:k-1
+                                B[bidx(i,j)] -= B[bidx(k,j)]*A[aidx(i,k)]
+                            end
+                        end
+                    end
+                end
+            else # !upper
+                @inbounds for j=1:n
+                    if !isone(alpha)
+                        @inbounds for i=1:m
+                            B[bidx(i,j)] *= alpha
+                        end
+                    end
+                    @inbounds for k=1:m
+                        if !iszero(B[bidx(k,j)])
+                            if nounit
+                                B[bidx(k,j)] /= A[aidx(k,k)]
+                            end
+                            @inbounds for i=k+1:m
+                                B[bidx(i,j)] -=  B[bidx(k,j)]*A[aidx(i,k)]
+                            end
+                        end
+                    end
+                end
+            end
+        else # transa=='t'
+            #  @Inbounds Form  B := alpha*inv( A**T )*B.
+            if upper
+                @inbounds for j=1:n
+                    @inbounds for i=1:m
+                        temp = alpha*B[bidx(i,j)]
+                        @inbounds for k=1:i-1
+                            temp -=  A[aidx(k,i)]*B[bidx(k,j)]
+                        end
+                        B[bidx(i,j)] = temp
+                        if nounit
+                            temp /= A[aidx(i,i)]
+                        end
+                        B[bidx(i,j)]=temp
+                    end
+                end
+            else #!upper
+                @inbounds for j=1:n
+                    @inbounds for i=m:-1:1
+                        temp = alpha*B[bidx(i,j)]
+                        @inbounds for k=i+1:m
+                            temp -= A[aidx(k,i)]*B[bidx(k,j)]
+                        end
+                        if nounit
+                            temp /= A[aidx(i,i)]
+                        end
+                        B[bidx(i,j)] = temp
+                    end
+                end
+            end
+        end
+    else # !lside
+        if transA== 'n'
+            #    @Inbounds Form  B := alpha*B*inv( A ).
+            if upper
+                @inbounds for j=1:n
+                    if !isone(alpha)
+                        @inbounds for i=1:m
+                            B[bidx(i,j)] = alpha*B[bidx(i,j)]
+                        end
+                    end
+                    @inbounds for k=1:j-1
+                        if !iszero(A[aidx(k,j)])
+                            @inbounds for i=1:m
+                                B[bidx(i,j)] -= A[aidx(k,j)]*B[bidx(i,k)]
+                            end
+                        end
+                    end
+                    if nounit
+                        temp=oneT/A[aidx(j,j)]
+                        @inbounds for i=1:m
+                            B[bidx(i,j)] *= temp
+                        end
+                    end
+                end
+            else #!upper
+                @inbounds for j=n:-1:1
+                    if !isone(alpha)
+                        @inbounds for i=1:m
+                            B[bidx(i,j)] *= alpha
+                        end
+                    end
+                    @inbounds for k=j+1:n
+                        if !iszero(A[aidx(k,j)])
+                            @inbounds for i=1:m
+                                B[bidx(i,j)] -= A[aidx(k,j)]*B[bidx(i,k)]
+                            end
+                        end
+                    end
+                    if nounit
+                        temp = oneT/A[aidx(j,j)]
+                        @inbounds for i=1:m
+                            B[bidx(i,j)] *= temp
+                        end
+                    end
+                end
+            end
+        else
+            # @Inbounds Form  B := alpha*B*inv( A**T ).
+            if upper
+                @inbounds for k=n:-1:1
+                    if nounit
+                        temp = oneT/A[aidx(k,k)]
+                        @inbounds for i=1:m
+                            B[bidx(i,k)] *= temp
+                        end
+                    end
+                    @inbounds for j=1:k-1
+                        if !iszero(A[aidx(j,k)])
+                            temp = A[aidx(j,k)]
+                            @inbounds for i=1:m
+                                B[bidx(i,j)] -= temp*B[bidx(i,k)]
+                            end
+                        end
+                    end
+                    if !isone(alpha)
+                        @inbounds for i=1:m
+                            B[bidx(i,k)] *= alpha
+                        end
+                    end
+                end
+            else #!upper
+                @inbounds for k=1:n
+                    if nounit
+                        temp = oneT/A[aidx(k,k)]
+                        @inbounds for i=1:m
+                            B[bidx(i,k)] *= temp
+                        end
+                    end
+                    @inbounds for j=k+1:n
+                        if !iszero(A[aidx(j,k)])
+                            temp = A[aidx(j,k)]
+                            @inbounds for i=1:m
+                                B[bidx(i,j)] -=  temp*B[bidx(i,k)]
+                            end
+                        end
+                    end
+                    if !isone(alpha)
+                        @inbounds for i=1:m
+                            B[bidx(i,k)] *= alpha
+                        end
+                    end
+                end
+            end
+        end
     end
 end
-
 
 
 #
 # Vector entry swap according to ipiv
 #
 function glaswp!(a,lda,k1,k2,ipiv)
-    for i=k1:k2
+    @inbounds for i=k1:k2
         ip=ipiv[i]
         if ip!=i
             temp=a[i]