Before:
```
generic003._test(N=[10,100,1000,10_000, 100_000])
n=10:
                                 Float64:  10.007 μs (113 allocations: 14.02 KiB)
      MultiFloats.MultiFloat{Float64, 1}:  19.844 μs (346 allocations: 49.84 KiB)
      MultiFloats.MultiFloat{Float64, 2}:  24.761 μs (378 allocations: 33.75 KiB)
   ForwardDiff.Dual{Float64, Float64, 1}:  24.175 μs (430 allocations: 38.52 KiB)

n=100:
                                 Float64:  85.403 μs (113 allocations: 82.17 KiB)
      MultiFloats.MultiFloat{Float64, 1}:  222.388 μs (3650 allocations: 336.25 KiB)
      MultiFloats.MultiFloat{Float64, 2}:  306.651 μs (4300 allocations: 527.41 KiB)
   ForwardDiff.Dual{Float64, Float64, 1}:  272.310 μs (4300 allocations: 669.56 KiB)

n=1000:
                                 Float64:  1.299 ms (132 allocations: 800.73 KiB)
      MultiFloats.MultiFloat{Float64, 1}:  2.956 ms (46817 allocations: 15.19 MiB)
      MultiFloats.MultiFloat{Float64, 2}:  3.462 ms (41737 allocations: 7.95 MiB)
   ForwardDiff.Dual{Float64, Float64, 1}:  3.126 ms (44044 allocations: 10.52 MiB)

n=10000:
                                 Float64:  38.397 ms (201 allocations: 9.45 MiB)
      MultiFloats.MultiFloat{Float64, 1}:  172.454 ms (969317 allocations: 1.10 GiB)
      MultiFloats.MultiFloat{Float64, 2}:  640.790 ms (974944 allocations: 1.12 GiB)
   ForwardDiff.Dual{Float64, Float64, 1}:  199.685 ms (966803 allocations: 1.10 GiB)

n=100000:
                                 Float64:  17.629 s (203 allocations: 279.91 MiB)
      MultiFloats.MultiFloat{Float64, 1}:  73.733 s (72530297 allocations: 137.63 GiB)
      MultiFloats.MultiFloat{Float64, 2}:  590.859 s (74948908 allocations: 139.45 GiB)
   ForwardDiff.Dual{Float64, Float64, 1}:  100.656 s (65286991 allocations: 120.63 GiB)
```

This PR:
```
generic003._test(N=[10,100,1000,10_000, 100_000])
n=10:
                                 Float64:  10.139 μs (113 allocations: 14.02 KiB)
      MultiFloats.MultiFloat{Float64, 1}:  8.184 μs (114 allocations: 15.03 KiB)
      MultiFloats.MultiFloat{Float64, 2}:  10.080 μs (114 allocations: 16.41 KiB)
   ForwardDiff.Dual{Float64, Float64, 1}:  9.360 μs (114 allocations: 17.14 KiB)
n=100:
                                 Float64:  82.238 μs (113 allocations: 82.17 KiB)
      MultiFloats.MultiFloat{Float64, 1}:  70.719 μs (114 allocations: 88.59 KiB)
      MultiFloats.MultiFloat{Float64, 2}:  88.428 μs (114 allocations: 102.77 KiB)
   ForwardDiff.Dual{Float64, Float64, 1}:  79.392 μs (114 allocations: 104.25 KiB)
n=1000:
                                 Float64:  1.318 ms (132 allocations: 800.73 KiB)
      MultiFloats.MultiFloat{Float64, 1}:  942.338 μs (134 allocations: 814.84 KiB)
      MultiFloats.MultiFloat{Float64, 2}:  1.145 ms (134 allocations: 952.12 KiB)
   ForwardDiff.Dual{Float64, Float64, 1}:  1.059 ms (137 allocations: 1012.48 KiB)
n=10000:
                                 Float64:  38.081 ms (201 allocations: 9.45 MiB)
      MultiFloats.MultiFloat{Float64, 1}:  30.215 ms (203 allocations: 9.82 MiB)
      MultiFloats.MultiFloat{Float64, 2}:  148.908 ms (203 allocations: 13.11 MiB)
   ForwardDiff.Dual{Float64, Float64, 1}:  54.052 ms (203 allocations: 12.93 MiB)
n=100000:
                                 Float64:  17.567 s (203 allocations: 279.91 MiB)
      MultiFloats.MultiFloat{Float64, 1}:  17.181 s (205 allocations: 290.91 MiB)
      MultiFloats.MultiFloat{Float64, 2}:  119.715 s (205 allocations: 496.59 MiB)
   ForwardDiff.Dual{Float64, Float64, 1}:  45.105 s (205 allocations: 487.82 MiB)
```

With this PR, the generic implementation is on par with the binary blas/lapack wrt. allocations. Before, their number was prohibitively large for any larger sized problem.

After all, I just rewrote the code from netlib. It is BSD licensed (see https://netlib.org/lapack), so there seems to be no problem including this here.

The results conform to the mantra regarding Julia performance: "In doubt, write your own loop" which worked quite well for me in several situations before.

What is puzzling is that the generic routines are faster than the standard ones (e.g. compare between Float64 and MultiFloat{Float64,1}). However I would not yet jump to conclusions here before seeing more tests with relevant problems.

I think this PR should be implemented before any announcement...