@PetrKryslUCSD 2022-09-27
Done.

On Tue, Sep 27, 2022, 10:40 AM Jürgen Fuhrmann ***@***.***>
wrote:

> Dear Petr, could you tag a release ? I would like to announce this when
> the better performing version becomes available.
>
> —
> Reply to this email directly, view it on GitHub
> <https://urldefense.com/v3/__https://github.com/PetrKryslUCSD/Sparspak.jl/pull/10*issuecomment-1259840692__;Iw!!Mih3wA!D0JSEZKYAtAbqObrLt7EU3ziEQOBGMC-69i70HwAfz1xJtKkbWEnjDd6TpFWie122YoSsrcSTap0QLI-IVk2K56_$>,
> or unsubscribe
> <https://urldefense.com/v3/__https://github.com/notifications/unsubscribe-auth/ACLGGWCI4TRVDQEHODZYNILWAMWRFANCNFSM6AAAAAAQUTFMJY__;!!Mih3wA!D0JSEZKYAtAbqObrLt7EU3ziEQOBGMC-69i70HwAfz1xJtKkbWEnjDd6TpFWie122YoSsrcSTap0QLI-IXrwaEd0$>
> .
> You are receiving this because you modified the open/close state.Message
> ID: ***@***.***>
>